requirements.txt
# ml
torch
transformers
scikit-learn
pandas
setfit
sentencepiece
accelerate
peft
datasets
openai
# running notebooks
jupyter
# creating pipelines
invoke
# infra
python-dotenv
